update on BAA

+++ app/Http/Controllers/Accounting/Transaction/FabController.php
@@ -17,6 +17,7 @@ use App\Models\TaxSetting;
 use App\Models\UnitType;
 use App\Models\User;
 use App\Service\Transaction\FabService;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -34,6 +35,7 @@ class FabController extends Controller
     private FabServiceCategory $fabServiceCategory;
     private SKL $skl;
     private UnitType $unitType;
+    private User $user;
 
     public function __construct()
     {
@@ -46,39 +48,49 @@ class FabController extends Controller
         $this->user = new User();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function index(): View
     {
+        $this->authorize('viewAny', Fab::class);
         return view('pages.transaction.fab.index');
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function data(): JsonResponse
     {
+        $this->authorize('view', Fab::class);
         return response()->json($this->fabService->data());
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function search(Request $request): JsonResponse
     {
+        $this->authorize('view', Fab::class);
         return response()->json($this->fabService->search($request));
     }
 
-
-    public function branchData(Request $request): JsonResponse
-    {
-        return response()->json($this->branch->getData($request));
-    }
-
-    public function filterByBranch(Branch $branch): JsonResponse
-    {
-        return response()->json($this->fabService->filterByBranch($branch->id, $this->perPage));
-    }
-
+    /**
+     * @throws AuthorizationException
+     */
     public function create(): View
     {
+        $this->authorize('create', Fab::class);
         return view('pages.transaction.fab.create');
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function contactData(Request $request): JsonResponse
     {
+        $this->authorize('create', Fab::class);
+        $this->authorize('update', Fab::class);
         $contact = $this->contact->getData($request);
         return response()->json($contact);
     }
@@ -89,30 +101,54 @@ class FabController extends Controller
         return response()->json($test);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getServicesCategoriesData(Request $request): JsonResponse
     {
+        $this->authorize('create', Fab::class);
+        $this->authorize('update', Fab::class);
         $services = $this->serviceCategory->getData($request);
         return response()->json($services);
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getSKL(Request $request): JsonResponse
     {
+        $this->authorize('create', Fab::class);
+        $this->authorize('update', Fab::class);
         return response()->json($this->skl->getData($request));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getUnitType(Request $request): JsonResponse
     {
+        $this->authorize('create', Fab::class);
+        $this->authorize('update', Fab::class);
         return response()->json($this->unitType->getData($request));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getUserData(Request $request): JsonResponse
     {
+        $this->authorize('create', Fab::class);
+        $this->authorize('update', Fab::class);
         return response()->json($this->user->getUser($request));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getSelectedUser(Fab $fab): JsonResponse
     {
+        $this->authorize('update', $fab);
         return response()->json($this->user->getSelectedData($fab->pic));
     }
 
@@ -121,6 +157,7 @@ class FabController extends Controller
      */
     public function store(FabRequest $request): JsonResponse
     {
+        $this->authorize('create', Fab::class);
         $this->fabService->store($request);
         return response()->json([
             'message' => 'Data berhasil disimpan',
@@ -151,8 +188,14 @@ class FabController extends Controller
         return strtoupper($newPTKey) . ' ' . ucwords(trim($newString));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function detail(Fab $fab): View
     {
+
+        $this->authorize('viewDetail', Fab::class);
+
         $fabHasServiceCategories = FabServiceCategory::with('service', 'unitType')
             ->where('fab_id', $fab->id)
             ->get();
@@ -187,26 +230,46 @@ class FabController extends Controller
         return view('pages.transaction.fab.view-file', compact('fab'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function edit(Fab $fab): View
     {
+        $this->authorize('update', $fab);
+
         return view('pages.transaction.fab.edit', compact('fab'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function selectedContact(Fab $fab): JsonResponse
     {
+        $this->authorize('update', $fab);
+
         $selectedContact = $this->contact->getSelectedData($fab->contact_id);
 
         return response()->json($selectedContact);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function selectedServices(Fab $fab): JsonResponse
     {
+        $this->authorize('update', $fab);
+
         $fabService = FabServiceCategory::where('fab_id', $fab->id)->get();
         return response()->json($fabService);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function selectedSKL(Fab $fab): JsonResponse
     {
+        $this->authorize('update', $fab);
+
         $fabSKL = FabHasSKL::where('fab_id', $fab->id)->get();
         return response()->json($fabSKL);
     }
@@ -216,6 +279,8 @@ class FabController extends Controller
      */
     public function update(FabRequest $request, Fab $fab): JsonResponse
     {
+        $this->authorize('update', $fab);
+
         $this->fabService->update($request, $fab);
 
         return response()->json([
@@ -228,6 +293,7 @@ class FabController extends Controller
      */
     public function confirm(Fab $fab): JsonResponse
     {
+        $this->authorize('confirm', $fab);
         $this->fabService->confirm($fab);
 
         return response()->json([
@@ -235,17 +301,24 @@ class FabController extends Controller
         ]);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(Fab $fab): JsonResponse
     {
+        $this->authorize('delete', $fab);
         $fab->delete();
-
         return response()->json([
             'message' => 'data berhasil dihapus',
         ]);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function exportPDF(Fab $fab): Response
     {
+        $this->authorize('print', $fab);
         $fabHasServiceCategories = FabServiceCategory::with('service', 'unitType')
             ->where('fab_id', $fab->id)
             ->get();
@@ -295,9 +368,26 @@ class FabController extends Controller
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function contractPDF(Fab $fab): Response
     {
-        $view = view('pages.transaction.fab.contract.index', compact('fab'));
+        $this->authorize('printContract', $fab);
+
+
+        $fabServiceCategories = FabServiceCategory::with('service')->where('fab_id', $fab->id)->get();
+
+        $serviceCategories = [];
+
+        foreach ($fabServiceCategories as $serviceCategory) {
+            $serviceCategories[] = $serviceCategory->service->name;
+        }
+
+        $serviceCategories = implode(', ', $serviceCategories);
+
+
+        $view = view('pages.transaction.fab.contract.index', compact('fab', 'serviceCategories', 'fabServiceCategories'));
         $pdf = Browsershot::html($view)
             ->setChromePath('/usr/bin/chromium')
             ->noSandbox()
